requirements.txt
asgiref==3.7.2
Django==4.2.2
django-cors-headers==4.0.0
django-environ==0.10.0
djangorestframework==3.14.0
psycopg==3.1.9
psycopg-binary==3.1.9
Pygments==2.15.1
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
